refactor(vibe3): merge resources module into environment
Move runtime_assets.py from resources/ to environment/ module.
This consolidation reduces module count and improves code organization.

Changes:
- Move src/vibe3/resources/runtime_assets.py → environment/
- Update environment/__init__.py to export runtime assets functions
- Update 6 import statements across agents/ and prompts/
- Remove src/vibe3/resources/ directory

Rationale:
- resources (66 lines) was too small for a top-level module
- Runtime asset resolution naturally belongs in environment module
- environment already manages worktree and session contexts
