+++ src/vibe3/environment/__init__.py
@@ -1,5 +1,10 @@
-"""Environment isolation modules (worktree, session)."""
+"""Environment isolation modules (worktree, session, runtime assets)."""
 
+from vibe3.environment.runtime_assets import (
+    resolve_prompt_config,
+    resolve_runtime_asset,
+    runtime_assets_root,
+)
 from vibe3.environment.session import (
     CodeagentSessionContext,
     SessionManager,
@@ -14,4 +19,7 @@ __all__ = [
     "TmuxSessionContext",
     "CodeagentSessionContext",
     "SessionManager",
+    "resolve_prompt_config",
+    "resolve_runtime_asset",
+    "runtime_assets_root",
 ]
