+++ src/vibe3/environment/runtime_assets.py
