+++ src/vibe3/prompts/builtin_providers.py
@@ -9,10 +9,10 @@ from typing import Any
 
 from loguru import logger
 
+from vibe3.environment.runtime_assets import resolve_runtime_asset
 from vibe3.prompts.exceptions import ProviderNotFoundError
 from vibe3.prompts.models import PromptVariableSource, VariableSourceKind
 from vibe3.prompts.provider_registry import ProviderRegistry
-from vibe3.resources.runtime_assets import resolve_runtime_asset
 
 
 def resolve_skill_content(
