+++ src/vibe3/prompts/manifest.py
@@ -10,6 +10,7 @@ from typing import Any
 import yaml
 from loguru import logger
 
+from vibe3.environment.runtime_assets import resolve_prompt_config
 from vibe3.prompts.models import (
     LoadedPromptRecipeDefinition,
     PromptRecipeKind,
@@ -18,7 +19,6 @@ from vibe3.prompts.models import (
     PromptVariableSource,
     VariableSourceKind,
 )
-from vibe3.resources.runtime_assets import resolve_prompt_config
 
 DEFAULT_PROMPT_RECIPES_PATH = Path("config/prompts/prompt-recipes.yaml")
