--- src/vibe3/resources/__init__.py
@@ -1,1 +1,0 @@
-"""Runtime resource path helpers."""
